requirements.txt
Django==3.2.12
django_environ==0.10.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.3.0
